rename types to np_types to avoid confusion with python core module types
