+++ neuralprophet/np_types.py
